Refactor game rendering responsibilities
